+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/RealmResultsImpl.kt
@@ -80,8 +80,6 @@ internal class RealmResultsImpl<E : BaseRealmObject> constructor(
                 clazz,
                 mediator,
                 queryPointer,
-                query,
-                *args
             )
         } catch (exception: Throwable) {
             throw CoreExceptionConverter.convertToPublicException(