+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/query/ObjectQuery.kt
@@ -32,8 +32,10 @@ import io.realm.kotlin.internal.interop.RealmCoreIndexOutOfBoundsException
 import io.realm.kotlin.internal.interop.RealmCoreInvalidQueryException
 import io.realm.kotlin.internal.interop.RealmCoreInvalidQueryStringException
 import io.realm.kotlin.internal.interop.RealmInterop
+import io.realm.kotlin.internal.interop.RealmPointer
 import io.realm.kotlin.internal.interop.RealmQueryPointer
 import io.realm.kotlin.internal.interop.RealmResultsPointer
+import io.realm.kotlin.internal.interop.RealmValue
 import io.realm.kotlin.notifications.ResultsChange
 import io.realm.kotlin.query.RealmQuery
 import io.realm.kotlin.query.RealmResults
@@ -51,22 +53,35 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
     private val classKey: ClassKey,
     private val clazz: KClass<E>,
     private val mediator: Mediator,
-    composedQueryPointer: RealmQueryPointer? = null,
-    private val filter: String,
-    private vararg val args: Any?
+    private val queryPointer: RealmQueryPointer,
 ) : RealmQuery<E>, InternalDeleteable, Thawable<Observable<RealmResultsImpl<E>, ResultsChange<E>>>, Flowable<ResultsChange<E>> {
-
-    internal val queryPointer: RealmQueryPointer = when {
-        composedQueryPointer != null -> composedQueryPointer
-        else -> parseQuery()
-    }
 
     private val resultsPointer: RealmResultsPointer by lazy {
         RealmInterop.realm_query_find_all(queryPointer)
     }
 
     internal constructor(
-        composedQueryPointer: RealmQueryPointer?,
+        realmReference: RealmReference,
+        key: ClassKey,
+        clazz: KClass<E>,
+        mediator: Mediator,
+        filter: String,
+        args: Array<out Any?>
+    ) : this(
+        realmReference,
+        key,
+        clazz,
+        mediator,
+        parseQuery(
+            realmReference.dbPointer,
+            key,
+            filter,
+            RealmValueArgumentConverter.convertArgs(args)
+        ),
+    )
+
+    internal constructor(
+        composedQueryPointer: RealmQueryPointer,
         objectQuery: ObjectQuery<E>
     ) : this(
         objectQuery.realmReference,
@@ -74,8 +89,6 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
         objectQuery.clazz,
         objectQuery.mediator,
         composedQueryPointer,
-        objectQuery.filter,
-        *objectQuery.args
     )
 
     override fun find(): RealmResults<E> =
@@ -171,38 +184,41 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
         find().asInternalDeleteable().delete()
     }
 
-    private fun parseQuery(): RealmQueryPointer = tryCatchCoreException {
-        RealmInterop.realm_query_parse(
-            realmReference.dbPointer,
-            classKey,
-            filter,
-            RealmValueArgumentConverter.convertArgs(args)
-        )
-    }
-
-    private fun tryCatchCoreException(block: () -> RealmQueryPointer): RealmQueryPointer = try {
-        block.invoke()
-    } catch (exception: Throwable) {
-        throw CoreExceptionConverter.convertToPublicException(
-            exception,
-            customMessage = "Invalid syntax in query: ${exception.message}"
-        ) { coreException: RealmCoreException ->
-            when (coreException) {
-                is RealmCoreInvalidQueryStringException ->
-                    IllegalArgumentException("Wrong query string: ${coreException.message}")
-                is RealmCoreInvalidQueryException ->
-                    IllegalArgumentException("Wrong query field provided or malformed syntax in query: ${coreException.message}")
-                is RealmCoreIndexOutOfBoundsException ->
-                    IllegalArgumentException("Have you specified all parameters in your query?: ${coreException.message}")
-                else -> {
-                    // Use default mapping
-                    null
-                }
-            }
-        }
-    }
-
     override fun description(): String {
         return RealmInterop.realm_query_get_description(queryPointer)
+    }
+
+    companion object {
+        private fun parseQuery(
+            dbPointer: RealmPointer,
+            classKey: ClassKey,
+            filter: String,
+            args: Array<RealmValue>
+        ): RealmQueryPointer = tryCatchCoreException {
+            RealmInterop.realm_query_parse(dbPointer, classKey, filter, args)
+        }
+
+        private fun tryCatchCoreException(block: () -> RealmQueryPointer): RealmQueryPointer =
+            try {
+                block.invoke()
+            } catch (exception: Throwable) {
+                throw CoreExceptionConverter.convertToPublicException(
+                    exception,
+                    customMessage = "Invalid syntax in query: ${exception.message}"
+                ) { coreException: RealmCoreException ->
+                    when (coreException) {
+                        is RealmCoreInvalidQueryStringException ->
+                            IllegalArgumentException("Wrong query string: ${coreException.message}")
+                        is RealmCoreInvalidQueryException ->
+                            IllegalArgumentException("Wrong query field provided or malformed syntax in query: ${coreException.message}")
+                        is RealmCoreIndexOutOfBoundsException ->
+                            IllegalArgumentException("Have you specified all parameters in your query?: ${coreException.message}")
+                        else -> {
+                            // Use default mapping
+                            null
+                        }
+                    }
+                }
+            }
     }
 }